Front_End/Pruebas/Final/requirements.txt
streamlit==1.20.0
pydeck==0.7.1
protobuf==3.19.5
streamlit_extras.app_logo
pandas==1.3.3
psycopg2-binary==2.9.1
streamlit-shadcn-ui==0.2.0